requirements.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography==37.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-ckeditor==6.4.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-paypal==2.0
et-xmlfile==1.1.0
h11==0.13.0
idna==3.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
MarkupPy==1.14
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.2
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tablib==3.2.1
tqdm @ file:///C:/ci/tqdm_1650636210717/work
trio==0.20.0
trio-websocket==0.9.2
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.6
wincertstore==0.2
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
